determining best score on every play
line 18 & line 141

if it's your first play, let's write to storage every time there's an increment to the roll counter.

The final roll count value will not be compared against any existing value as it will be the benchmark for future plays.

line 141 sets "numberofRolls" to 0 and "isFirstPlay" to false, then every play after will execute the else block, which reads from storage and stores the record in "lastnumberOfRolls", which will be compared against the new "numberofRolls"